requirements.txt
numpy
pandas
torch
scikit-learn
tqdm
pillow
pyyaml
